+++ esmvaltool/diag_scripts/climate_patterns/cp_plotting.py
